refactor: extract adventuresList and cruisesList into components
